refactor: split keyring-sdk + add /v2 exports
