+++ packages/keyring-eth-ledger-bridge/src/v2/ledger-keyring.ts
@@ -13,7 +13,7 @@ import {
   type Keyring,
   KeyringType,
 } from '@metamask/keyring-api/v2';
-import { EthKeyringWrapper } from '@metamask/keyring-sdk';
+import { EthKeyringWrapper } from '@metamask/keyring-sdk/v2';
 import type { AccountId, EthKeyring } from '@metamask/keyring-utils';
 import { add0x, getChecksumAddress, type Hex } from '@metamask/utils';
