+++ packages/keyring-sdk/src/eth/index.ts
@@ -1,2 +1,1 @@
 export * from './account-id';
-export * from './eth-keyring-wrapper';
