+++ packages/keyring-sdk/src/v2/eth/index.ts
@@ -1,0 +1,1 @@
+export * from './eth-keyring-wrapper';
