+++ packages/keyring-sdk/src/v2/index.ts
@@ -1,0 +1,2 @@
+export * from './keyring-wrapper';
+export * from './eth';
